added clear form button and implemented local storage hook in other contributors component
