+++ src/components/Form/ClearButton.tsx
@@ -1,7 +1,9 @@
-import useLocalStorageState from 'use-local-storage-state';
+import { Button } from '@nextui-org/react';
+import FormStorage from '@/utils/FormStorage';
 
 function ClearButton() {
-
+  const { form, setForm, removeItem } = FormStorage();
+  return (<Button auto onPress={removeItem}>Clear Form</Button>)
 }
 
 export default ClearButton;
