+++ src/components/Form/SourceDropdown.tsx
@@ -16,40 +16,38 @@ function SourceDropdown() {
   );
 
   return (
-    <div>
-      <Dropdown>
-        <Dropdown.Button light css={{ tt: 'capitalize' }}>
-          {selected ? selected : 'Source'}
-        </Dropdown.Button>
-        <Dropdown.Menu
-          aria-label='Sources'
-          variant='light'
-          disallowEmptySelection
-          selectionMode='single'
-          selectedKeys={selected}
-          onSelectionChange={setSelectedValue}
-        >
-          <Dropdown.Item key='website'>
-            <div className='flex items-center'>
-              <LayoutTextWindow className='h-6 w-auto mr-2' />
-              Website
-            </div>
-          </Dropdown.Item>
-          <Dropdown.Item key='book'>
-            <div className='flex items-center'>
-              <Book className='h-6 w-auto mr-2' />
-              Book
-            </div>
-          </Dropdown.Item>
-          <Dropdown.Item key='journal'>
-            <div className='flex items-center'>
-              <Journal className='h-6 w-auto mr-2' />
-              Journal
-            </div>
-          </Dropdown.Item>
-        </Dropdown.Menu>
-      </Dropdown>
-    </div>
+    <Dropdown>
+      <Dropdown.Button css={{ tt: 'capitalize'}}>
+        {selected ? selected : 'Source'}
+      </Dropdown.Button>
+      <Dropdown.Menu
+        aria-label='Sources'
+        variant='light'
+        disallowEmptySelection
+        selectionMode='single'
+        selectedKeys={selected}
+        onSelectionChange={setSelectedValue}
+      >
+        <Dropdown.Item key='website'>
+          <div className='flex items-center'>
+            <LayoutTextWindow className='h-6 w-auto mr-2' />
+            Website
+          </div>
+        </Dropdown.Item>
+        <Dropdown.Item key='book'>
+          <div className='flex items-center'>
+            <Book className='h-6 w-auto mr-2' />
+            Book
+          </div>
+        </Dropdown.Item>
+        <Dropdown.Item key='journal'>
+          <div className='flex items-center'>
+            <Journal className='h-6 w-auto mr-2' />
+            Journal
+          </div>
+        </Dropdown.Item>
+      </Dropdown.Menu>
+    </Dropdown>
   );
 }
