--- src/utils/citationStorage.ts
@@ -1,5 +1,0 @@
-import useLocalStorageState from 'use-local-storage-state';
-
-export function initializeStorage() {
-  
-}
